implement the pose overlay as a class in player_widget/overlays instead of directly in FrameWidget
