+++ src/jabs/ui/player_widget/frame_widget.py
@@ -10,17 +10,13 @@ from jabs.ui.colors import (
     BACKGROUND_COLOR,
     BEHAVIOR_COLOR,
     INACTIVE_ID_COLOR,
-    KEYPOINT_COLOR_MAP,
     NOT_BEHAVIOR_COLOR,
 )
-from jabs.utils.pose_util import gen_line_fragments
 
 _BEHAVIOR_LABEL_OUTLINE_COLOR = QtGui.QColor(255, 255, 255)
 _FONT_SIZE = 16  # size of the font used for identity labels
 _BEHAVIOR_LABEL_SIZE = 10  # size of the behavior label square
 _GAP = 5  # gap between identity label and behavior label
-_LINE_SEGMENT_COLOR = QtGui.QColor(255, 255, 255, 128)  # color for the pose line segments
-_KEYPOINT_SIZE = 3  # size of the keypoint circles
 _SELECTED_INDICATOR_SIZE = (
     2  # size of the circle drawn to indicate selected mouse when ID labels are hidden
 )
@@ -270,11 +266,6 @@ class FrameWidget(QtWidgets.QLabel):
             self._scaled_pix_width = pix.width()
             self._scaled_pix_height = pix.height()
 
-            if self._pose_overlay_mode == self.PoseOverlayMode.ALL:
-                self._overlay_pose(painter, all_identities=True)
-            elif self._pose_overlay_mode == self.PoseOverlayMode.ACTIVE_IDENTITY:
-                self._overlay_pose(painter, all_identities=False)
-
             if self._id_overlay_mode in (
                 self.IdentityOverlayMode.CENTROID,
                 self.IdentityOverlayMode.MINIMAL,
@@ -371,76 +362,3 @@ class FrameWidget(QtWidgets.QLabel):
             painter.setBrush(prediction_color)
             painter.setPen(_BEHAVIOR_LABEL_OUTLINE_COLOR)
             painter.drawRect(behavior_x, behavior_y, _BEHAVIOR_LABEL_SIZE, _BEHAVIOR_LABEL_SIZE)
-
-    def _overlay_pose(self, painter: QtGui.QPainter, all_identities: bool = False) -> None:
-        """Overlay pose estimation on the current frame.
-
-        This method draws the pose estimation skeletons on the frame. If `all_identities` is True,
-        it will draw all identities; otherwise, it will only draw the active identity with a more prominent color.
-
-        Args:
-            painter (QtGui.QPainter): The painter used to draw on the frame.
-            all_identities (bool): If True, draw all identities; if False, only draw the active identity.
-        """
-        if self._pose is None:
-            return
-
-        # draw the pose estimation skeletons
-        for identity in self._pose.identities:
-            if not all_identities and identity != self._active_identity:
-                continue
-
-            points, mask = self._pose.get_points(self._frame_number, identity)
-
-            if points is None or mask is None:
-                continue
-
-            # Adjust alpha for non-active identities
-            if identity != self._active_identity:
-                line_color = QtGui.QColor(_LINE_SEGMENT_COLOR)
-                line_color.setAlpha(line_color.alpha() // 3)  # More translucent
-            else:
-                line_color = _LINE_SEGMENT_COLOR
-
-            pen = QtGui.QPen(line_color)
-            pen.setWidth(3)
-            painter.setPen(pen)
-
-            for seg in gen_line_fragments(
-                self._pose.get_connected_segments(), np.flatnonzero(mask == 0)
-            ):
-                segment_points = [self.image_to_widget_coords(p[0], p[1]) for p in points[seg]]
-
-                # draw lines
-                if len(segment_points) >= 2:
-                    for i in range(len(segment_points) - 1):
-                        painter.drawLine(
-                            segment_points[i][0],
-                            segment_points[i][1],
-                            segment_points[i + 1][0],
-                            segment_points[i + 1][1],
-                        )
-
-            # draw points at each keypoint of the pose (if it exists at this frame)
-            painter.setPen(QtCore.Qt.PenStyle.NoPen)
-            for keypoint in PoseEstimation.KeypointIndex:
-                point_index = keypoint.value
-                if mask[point_index]:
-                    widget_x, widget_y = self.image_to_widget_coords(
-                        points[point_index][0], points[point_index][1]
-                    )
-
-                    # Use the color map to get the color for the keypoint
-                    # and make it translucent if it's not the active identity
-                    color = KEYPOINT_COLOR_MAP[keypoint]
-                    if identity != self._active_identity:
-                        # Make keypoints translucent for non-active identities
-                        translucent_color = QtGui.QColor(color)
-                        translucent_color.setAlpha(96)
-                        painter.setBrush(translucent_color)
-                    else:
-                        painter.setBrush(color)
-
-                    painter.drawEllipse(
-                        QtCore.QPoint(widget_x, widget_y), _KEYPOINT_SIZE, _KEYPOINT_SIZE
-                    )
